Fix NULL to int cast warnings
Neither of these two values needed to be set to `NULL`. One was only
using `NULL` as a means of providing a value which could never be
encountered from a real measurement (for which `-1` is suficient), and
the other was using it to produce an integer (for which it would have
been cast to `0` anyway).